Add a Vagrantfile for the project and fix building from scratch
This commit creates a Vagrantfile in the tools directory that
specifies an appropriate VM for the project. Additionally it fixes
a couple of Makefile problems that were preventing me from building
the project from scratch. The first problem was an incompatibility
with case-insensative filesystems. (The Makefile in the src directory
wrote a text file named "PLATFORM", which obviously conflicted with
the "platform" directory. The other problem was an erroneously
commented out line in the Makefile in the src/components/lib
directory. At least on my system, the project cannot be built
without that line.